inlining gpusatutils fixing 'mult definitions' problem
